overlay: SimpleDeviceConfig: Don't pin camera app to save memory
The default camera app can be sizable in some cases, e.g. Google Camera.
The system will only pin up to the first 80 MiB of the APK file, as well
as the first 80 MiB of its odex. There are several problems with this:

  - We could easily end up with 160 MiB of camera app files pinned,
    which is a somewhat tall order with the amount of RAM on most of our
    targets
  - The data that gets pinned may not be the most critical data for
    launching the camera

Disable pinning of the camera app to save RAM on target devices.

Change-Id: Ic22234b462414efe3086ee69df9b60ed751f528d
Signed-off-by: Tiktodz <dotkit@electrowizard.me>